# frontend/openapi/spec.yaml
openapi: 3.0.3
info:
  title: LKFL Platform API
  version: 0.1.0
  description: White-label multi-tenant платформа корпоративных льгот
paths:
...
  /admin/engagements/categories:
    post:
      summary: Создать категорию
      operationId: createCategory
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementCategoryResponse'
  /admin/engagements/categories/{id}:
    put:
      summary: Обновить категорию
      operationId: updateCategory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '200':
          description: Категория обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementCategoryResponse'
    delete:
      summary: Удалить категорию
      operationId: deleteCategory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Категория удалена
...
  /admin/engagements/types/{id}/offers:
    post:
      summary: Создать оффер для энгейджмента
      operationId: createOffer
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferRequest'
      responses:
        '201':
          description: Оффер создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementOfferResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EngagementCategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
        sort_order:
          type: integer
      required: [id, slug, name, sort_order]
    EngagementOfferResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        cost_cents:
          type: integer
        sort_order:
          type: integer
      required: [id, name, cost_cents, sort_order]
...
    CreateCategoryRequest:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
        sort_order:
          type: integer
      required: [slug, name, icon, sort_order]
...
    CreateOfferRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        cost_cents:
          type: integer
        sort_order:
          type: integer
      required: [name, cost_cents, sort_order]